relay/requirements.txt
Flask==1.1.1
waitress==1.4.3
guppy3
hdfs